+++ frontend/src/theme/index.ts
@@ -1,6 +1,6 @@
 import { extendTheme } from '@chakra-ui/react'
 
-import { colours } from './colours'
+import { colours } from './foundations/colours'
 import { components } from './components'
 import { textStyles } from './textStyles'
